+++ modules/app/recording/recording.tsx
@@ -28,6 +28,7 @@ import {
 import Modal from "react-modal";
 import MedicalTranscriptionAPI from "@/services/api/genesis-api.service";
 import recording_styles from "./styles/recording.module.css";
+import DeleteConfirmation from "@/resources/containers/delete-confirmation";
 
 export const getStaticProps: GetStaticProps = getStaticPropsWithTranslations;
 Modal.setAppElement("#__next");
@@ -51,36 +52,33 @@ const Recording = () => {
   }, [audioUrl]);
 
   const togglePlayPause = () => {
-    if (audioUrl && audioRef.current) {
-      if (isPlaying) {
-        audioRef.current.pause();
-      } else {
-        audioRef.current.play();
-      }
-      setIsPlaying(!isPlaying);
-    }
+    if (!(audioUrl && audioRef.current)) return;
+
+    audioRef.current[isPlaying ? "pause" : "play"]();
+    setIsPlaying(!isPlaying);
   };
 
   const handleTimeUpdate = () => {
-    if (audioRef.current) {
-      setCurrentTime(audioRef.current.currentTime);
-      updateSeekBarProgress();
-    }
+    if (!audioRef.current) return;
+
+    setCurrentTime(audioRef.current.currentTime);
+    updateSeekBarProgress();
   };
 
   const handleSeek = (e: React.ChangeEvent<HTMLInputElement>) => {
+    if (!audioRef.current) return;
+
     const time = Number(e.target.value);
     setCurrentTime(time);
-    if (audioRef.current) {
-      audioRef.current.currentTime = time;
-      updateSeekBarProgress();
-    }
+
+    audioRef.current.currentTime = time;
+    updateSeekBarProgress();
   };
 
   const handleLoadedMetadata = () => {
-    if (audioRef.current) {
-      setDuration(audioRef.current.duration);
-    }
+    if (!audioRef.current) return;
+
+    setDuration(audioRef.current.duration);
   };
 
   const formatTime = (time: number) => {
@@ -90,30 +88,30 @@ const Recording = () => {
   };
 
   const handleSkip = (seconds: number) => {
-    if (audioRef.current) {
-      audioRef.current.currentTime += seconds;
-    }
+    if (!audioRef.current) return;
+
+    audioRef.current.currentTime += seconds;
   };
 
   const updateSeekBarProgress = () => {
-    if (audioRef.current) {
-      const progressPercent = (currentTime / duration) * 100;
-      const seekBar = document.querySelector(
-        `.${recording_styles.seekBar}`,
-      ) as HTMLInputElement;
-      if (seekBar) {
-        seekBar.style.setProperty("--seek-before-width", `${progressPercent}%`);
-      }
+    if (!audioRef.current) return;
+
+    const progressPercent = (currentTime / duration) * 100;
+    const seekBar = document.querySelector(
+      `.${recording_styles.seekBar}`,
+    ) as HTMLInputElement;
+    if (seekBar) {
+      seekBar.style.setProperty("--seek-before-width", `${progressPercent}%`);
     }
   };
 
   const handleStop = () => {
-    if (audioRef.current) {
-      audioRef.current.pause();
-      audioRef.current.currentTime = 0;
-      setIsPlaying(false);
-      updateSeekBarProgress();
-    }
+    if (!audioRef.current) return;
+
+    audioRef.current.pause();
+    audioRef.current.currentTime = 0;
+    setIsPlaying(false);
+    updateSeekBarProgress();
   };
 
   const handleEnded = () => {
@@ -123,41 +121,41 @@ const Recording = () => {
   };
 
   const handleSave = () => {
-    if (audioRef.current && audioUrl) {
-      const anchor = document.createElement("a");
-      anchor.href = audioUrl as string;
+    if (!(audioUrl && audioRef.current)) return;
 
-      anchor.download = typeof audioUrl === "string" ? audioUrl : audioUrl[0];
+    const anchor = document.createElement("a");
+    anchor.href = audioUrl as string;
 
-      document.body.appendChild(anchor);
-      anchor.click();
-      document.body.removeChild(anchor);
-    }
+    anchor.download = typeof audioUrl === "string" ? audioUrl : audioUrl[0];
+
+    document.body.appendChild(anchor);
+    anchor.click();
+    document.body.removeChild(anchor);
   };
 
   const handleSubmit = async () => {
+    if (!(audioUrl && audioRef.current)) return;
+
     setIsLoading(true);
 
-    if (audioRef.current && audioUrl) {
-      const response = await fetch(audioUrl as string);
-      const blob = await response.blob();
-      const file = new File([blob], "audio.mp3", { type: "audio/mpeg" });
+    const response = await fetch(audioUrl as string);
+    const blob = await response.blob();
+    const file = new File([blob], "audio.mp3", { type: "audio/mpeg" });
 
-      const api_response =
-        await MedicalTranscriptionAPI.processAudioAndGenerateReport(file);
+    const api_response =
+      await MedicalTranscriptionAPI.processAudioAndGenerateReport(file);
 
-      if (api_response) {
-        router.push({
-          pathname: "/app/report",
-          query: {
-            report: api_response.report,
-            transcription: api_response.transcription,
-          },
-        });
-      } else {
-        messageHandler.handleError("Failed to generate report");
-        setIsLoading(false);
-      }
+    if (api_response) {
+      router.push({
+        pathname: "/app/report",
+        query: {
+          report: api_response.report,
+          transcription: api_response.transcription,
+        },
+      });
+    } else {
+      messageHandler.handleError("Failed to generate report");
+      setIsLoading(false);
     }
   };
 
@@ -292,30 +290,12 @@ const Recording = () => {
             {t("Submit")}
           </button>
         </div>
-        <Modal
-          isOpen={isDeleteModalOpen}
-          onRequestClose={cancelDelete}
-          className={recording_styles.modal}
-          overlayClassName={recording_styles.overlay}
-        >
-          <h2>{t("Confirm Delete")}</h2>
-          <p>{t("Are you sure you want to delete this recording?")}</p>
-          <div className={recording_styles.modalButtons}>
-            <button
-              onClick={confirmDelete}
-              className={recording_styles.deleteButton}
-            >
-              {t("Delete")}
-            </button>
-            <button
-              onClick={cancelDelete}
-              className={recording_styles.cancelButton}
-            >
-              {t("Cancel")}
-            </button>
-          </div>
-        </Modal>
       </main>
+      <DeleteConfirmation
+        isOpen={isDeleteModalOpen}
+        onRequestClose={cancelDelete}
+        onConfirm={confirmDelete}
+      />
     </>
   );
 };
