+++ modules/app/recording/styles/recording.module.css
@@ -172,74 +172,3 @@
 .submitButton:hover {
   background-color: #4DBCAA;
 }
-
-.modal {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: white;
-  padding: 2rem;
-  border-radius: 8px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  max-width: 400px;
-  width: 100%;
-  color: black;
-  font-family: 'Montserrat', sans-serif;
-}
-
-.modal p,
-.modal h2 {
-  padding-bottom: 0.5rem;
-}
-
-.modal p {
-  font-size: 0.8rem;
-}
-
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.modalButtons {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 1rem;
-}
-
-.deleteButton, .cancelButton {
-  padding: 0.5rem 1rem;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-weight: bold;
-  font-size: 0.8rem;
-  transition: background-color 0.3s;
-}
-
-.deleteButton {
-  background-color: #DF4949;
-  color: white;
-  margin-right: 0.5rem;
-}
-
-.deleteButton:hover {
-  background-color: #ff3333;
-}
-
-.cancelButton {
-  background-color: #f0f0f0;
-  color: #333;
-}
-
-.cancelButton:hover {
-  background-color: #e0e0e0;
-}
