+++ modules/auth/supabase/confirm-reset-password.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { useTranslations } from "next-intl";
-import ConfirmResetPasswordForm from "@/core/forms/confirm-reset-password-form";
+import ConfirmResetPasswordForm from "./forms/confirm-reset-password.form";
 import { getStaticPropsWithTranslations } from "@/modules/lang/props";
 import { GetStaticProps } from "next";
 import { GlobalCore } from "@/core/module/module.types";
