+++ modules/auth/supabase/forms/confirm-reset-password.form.tsx
@@ -1,39 +1,42 @@
 import React from "react";
-import { useRouter } from "next/router";
 import { useForm } from "react-hook-form";
 import { useTranslations } from "next-intl";
-import { Provider } from "@supabase/supabase-js";
 import { yupResolver } from "@hookform/resolvers/yup";
 import AuthService from "@/services/auth/auth-supabase.service";
 import form_style from "./styles/form.module.css";
-import { reset_password_schema } from "./schemas/auth-schema";
+import { confirm_reset_password_schema } from "./auth.schema";
 import messageHandler from "@/core/message-handler";
 import { getStaticPropsWithTranslations } from "@/modules/lang/props";
 import { GetStaticProps } from "next";
 import Input from "@/resources/inputs/input";
 import AuthButton from "@/resources/containers/auth-button";
-import OAuthButton from "@/resources/containers/oauth-button";
 
 export const getStaticProps: GetStaticProps = getStaticPropsWithTranslations;
 
-const ResetPasswordForm = () => {
+interface confirmResetPasswordInterface {
+  onSuccess: () => void;
+}
+
+const ConfirmResetPasswordForm: React.FC<confirmResetPasswordInterface> = ({
+  onSuccess,
+}) => {
   const t = useTranslations("");
-  const router = useRouter();
   const {
     register,
     handleSubmit,
     formState: { errors },
   } = useForm({
-    resolver: yupResolver(reset_password_schema),
+    resolver: yupResolver(confirm_reset_password_schema),
   });
 
   const onSubmit = async (data: any) => {
-    const response = await AuthService.resetPassword(data.email);
+    const response = await AuthService.updateUser(
+      undefined,
+      data.password as string,
+    );
     if (response) {
-      messageHandler.handleSuccess(
-        t("Password reset link has been sent to your email"),
-      );
-      router.push("/auth/login");
+      messageHandler.handleSuccess(t("Successfully updated password"));
+      onSuccess();
     }
   };
 
@@ -42,10 +45,14 @@ const ResetPasswordForm = () => {
       onSubmit={handleSubmit(onSubmit)}
       className={form_style.formContainer}
     >
-      <Input register={register} errors={errors} action="reset-password" />
+      <Input
+        register={register}
+        errors={errors}
+        action="confirm-reset-password"
+      />
       <AuthButton action="Reset Password" />
     </form>
   );
 };
 
-export default ResetPasswordForm;
+export default ConfirmResetPasswordForm;
