+++ modules/auth/supabase/forms/reset-password.form.tsx
@@ -1,10 +1,11 @@
 import React from "react";
+import { useRouter } from "next/router";
 import { useForm } from "react-hook-form";
 import { useTranslations } from "next-intl";
 import { yupResolver } from "@hookform/resolvers/yup";
 import AuthService from "@/services/auth/auth-supabase.service";
 import form_style from "./styles/form.module.css";
-import { confirm_reset_password_schema } from "./schemas/auth-schema";
+import { reset_password_schema } from "./auth.schema";
 import messageHandler from "@/core/message-handler";
 import { getStaticPropsWithTranslations } from "@/modules/lang/props";
 import { GetStaticProps } from "next";
@@ -13,30 +14,24 @@ import AuthButton from "@/resources/containers/auth-button";
 
 export const getStaticProps: GetStaticProps = getStaticPropsWithTranslations;
 
-interface confirmResetPasswordInterface {
-  onSuccess: () => void;
-}
-
-const ConfirmResetPasswordForm: React.FC<confirmResetPasswordInterface> = ({
-  onSuccess,
-}) => {
+const ResetPasswordForm = () => {
   const t = useTranslations("");
+  const router = useRouter();
   const {
     register,
     handleSubmit,
     formState: { errors },
   } = useForm({
-    resolver: yupResolver(confirm_reset_password_schema),
+    resolver: yupResolver(reset_password_schema),
   });
 
   const onSubmit = async (data: any) => {
-    const response = await AuthService.updateUser(
-      undefined,
-      data.password as string,
-    );
+    const response = await AuthService.resetPassword(data.email);
     if (response) {
-      messageHandler.handleSuccess(t("Successfully updated password"));
-      onSuccess();
+      messageHandler.handleSuccess(
+        t("Password reset link has been sent to your email"),
+      );
+      router.push("/auth/login");
     }
   };
 
@@ -45,14 +40,10 @@ const ConfirmResetPasswordForm: React.FC<confirmResetPasswordInterface> = ({
       onSubmit={handleSubmit(onSubmit)}
       className={form_style.formContainer}
     >
-      <Input
-        register={register}
-        errors={errors}
-        action="confirm-reset-password"
-      />
+      <Input register={register} errors={errors} action="reset-password" />
       <AuthButton action="Reset Password" />
     </form>
   );
 };
 
-export default ConfirmResetPasswordForm;
+export default ResetPasswordForm;
